initial weighed item impl

+++ app/src/main/java/com/paywithclerc/paywithclerc/service/UtilityService.kt
@@ -8,7 +8,7 @@ object UtilityService {
     /**
      * Calculate total cost for items in cart
      */
-    fun getTotalCost(items: List<Product>, quantities: List<Int>): Double {
+    fun getTotalCost(items: List<Product>, quantities: List<Double>): Double {
         var totalPrice = 0.0
         for (index in 0 until items.size) {
             totalPrice += items[index].cost * quantities[index]
